requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pytesseract==0.3.13
Pillow==11.1.0
opencv-python-headless==4.10.0.84
PyMuPDF==1.25.3
sentence-transformers==3.3.1
numpy==2.2.1
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1
